Fix MSVC warnings

+++ lib/src/Http2Transport.h
@@ -146,8 +146,8 @@ class Http2Transport : public HttpTransport
     size_t avaliableWindowSize = 0;
 
     // Configuration settings
-    const size_t windowIncreaseThreshold = 32768;
-    const size_t windowIncreaseSize = 10 * 1024 * 1024;  // 10 MiB
+    const uint32_t windowIncreaseThreshold = 32768;
+    const uint32_t windowIncreaseSize = 10 * 1024 * 1024;  // 10 MiB
 
     // Set after server settings are received
     bool serverSettingsReceived = false;
